requirements.txt
absl-py==2.3.0
alembic==1.14.0
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.0
flask-cors==5.0.1
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.58.4
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.73.0
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
Mako==1.3.8
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
narwhals==1.42.1
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==24.2
pandas==2.3.0
pillow==11.2.1
pip==25.0
protobuf==5.29.5
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
RapidFuzz==3.13.0
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.37
streamlit==1.45.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.5.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2